+++ src/main/resources/templates/write/header.html
@@ -132,8 +132,8 @@
 						</h1>
 					</div>
 					<div class="top-inner-third">
-						<a href="#" class="inner-third-login loginservice">저장</a><!--글쓰기-->
-						<a href="#" class="inner-third-login logoutservice">저장</a><!--로그인-->
+<!-- 						<a href="#" class="inner-third-login loginservice">저장</a>글쓰기 -->
+<!-- 						<a href="#" class="inner-third-login logoutservice">저장</a>로그인 -->
 					</div>
 					
 <!-- 					<div class="top-inner-second"> -->
